Add Repeatable Text Field [Publisher] (#413)
* check in working settings and pub for repeatable lists

* Poc/repeatable lists use state (#382)

* chore: formatting change only

* chore: remove onChanged for repeatable

Repeatable fields don't need to be unique so there's
no extra logic to run here.

* refactor: use useState instead of react-hook-form

The publisher form doesn't currently make use of
React Hook Form, so introducing it here may
complicate things more than we need at present.
(It introduces two ways to manage form inputs:
the vanilla HTML inputs we're currently using,
and those managed by React Hook Form.)

* refactor: make field uncontrolled


* add: condition for repeatable lists of objects. Check types in future fields within .

* add: condition for repeatable lists of objects. Check types in future fields within .

* adds text fields and graphql logic

* removes field_repeatable property from sanitize_field and check for array instead

* fixes radio items in text field settings

* adds text changes for repeatables

* merges main and adjust textfield for new title changes

* check for isrepeatable set in reset responses

* adds test for repeater field settings

* adjust test docs

* adds repeater unit test

* adds rest test

* splits text out to its own component

* adds multiline text area condition for multi line repeaters

* adds description to multi line fields

* adds text field table styling

* full width the table

* styles text repeater publisher experience

* styles input and textarea repeater fields

* adds full width to inputs in a repeater

* makes inputs full width

* respects required fields in multiple choice fields

* respects min/max of characters in text repeaters

* adds required validation only for first field in a list of fields

* removes blank text repeat entries on save

* adjusts add button wording for repeatable text

* adds a switch for disabling titles when repeater is active and the reverse

* Update includes/publisher/js/src/components/Field.jsx


* Update includes/publisher/js/src/components/Text.jsx


* Update includes/publisher/lib/field-functions.php


* Update includes/publisher/scss/_text-field.scss


* fix linting issues

* adjusts wait time for e2e taxomomy test

* fixes bad merge conflict
