+++ includes/publisher/scss/_text-field.scss
@@ -1,0 +1,132 @@
+@import "../../../includes/shared-assets/scss/_variables";
+
+.text {
+	.text-table {
+		flex-flow: row wrap;
+		flex: 1 100%;
+	}
+
+	.remove-container,
+	.add-container {
+		align-items: center;
+	}
+
+	.repeater-input {
+		align-items: center !important;
+	}
+
+	button.remove-item {
+		max-height: 30px;
+		padding: 0px 11px;
+	}
+
+	.repeater-input {
+		textarea,
+		.repeater-input input {
+			width: 100% !important;
+			margin: 0px !important;
+		}
+	}
+
+	.repeater-input-container input[type="text"] {
+		margin: 0px !important;
+		width: 100% !important;
+	}
+
+	svg.add.add-small.add-small-button {
+		color: #7e5cef;
+	}
+	.add-option a {
+		font-weight: bold;
+		color: #7e5cef;
+		line-height: 23px;
+	}
+
+	.add-option a span {
+		vertical-align: top;
+		margin-top: 11px;
+	}
+
+	.add-option textarea {
+		margin: 0;
+	}
+
+	.add-option {
+		height: 70px;
+		margin-left: 5px;
+	}
+
+	table {
+		border: solid 1px $color-gray-disabled-border;
+		border-spacing: 0;
+	}
+
+	table tr {
+		border-bottom: solid 1px $color-gray-disabled-border;
+	}
+
+	table th {
+		text-transform: capitalize;
+		padding-top: 14px;
+		padding-bottom: 14px;
+	}
+
+	table td {
+		padding-top: 0;
+		padding-bottom: 0;
+		border: solid 1px $color-gray-disabled-border;
+	}
+
+	table td.checkbox,
+	table td.info {
+		text-align: center;
+		vertical-align: middle;
+	}
+
+	table td.info {
+		&:hover {
+			cursor: pointer;
+		}
+
+		button {
+			background: none;
+			border: none;
+			margin-left: -3px;
+			padding: 0;
+
+			svg {
+				padding-top: 6px;
+			}
+		}
+	}
+
+	table tr.unselectable {
+		td {
+			border-bottom: 1px solid lighten($color-gray-disabled-border, 8);
+			box-shadow: none;
+		}
+
+		label {
+			color: $color-gray-light;
+			cursor: default;
+		}
+	}
+
+	table td label,
+	table tr.empty td span {
+		display: inline-block;
+		padding-top: 20px;
+		padding-bottom: 20px;
+		width: 100%;
+	}
+
+	table input[type="checkbox"] {
+		border: 1px solid #6b757b;
+		border-radius: 4px;
+		transform: scale(1.5);
+	}
+
+	table input[type="radio"] {
+		transform: scale(1.5);
+	}
+}
