+++ includes/publisher/scss/index.scss
@@ -13,6 +13,7 @@
 	@import "bootstrap/scss/utilities/_api";
 	@import "bootstrap/scss/_buttons";
 	@import "./relationship-modal";
+	@import "./text-field";
 	@import "../../settings/scss/_app-card";
 	@import "./table";
 	@import "./tooltip";
