+++ tests/acceptance/CreateTaxonomyCest.php
@@ -31,7 +31,7 @@ class CreateTaxonomyCest {
 		$i->fillField( [ 'name' => 'plural' ], 'Breeds' );
 		$i->click( '.checklist .checkbox' ); // The “goose” model.
 		$i->click( '.card-content button.primary' );
-		$i->wait( 1 );
+		$i->wait( 2 );
 		$i->see( 'taxonomy was created', '#success' );
 		$i->see( 'Breeds', '.taxonomy-list' );
 		$i->see( 'goose', '.taxonomy-list' );
